requirements.txt
requests
aiohttp
asyncio
pyyaml
python-telegram-bot
